ReFactorizacion del codigo
