+++ lib/widgets/widgets.dart
@@ -1,3 +1,9 @@
+export 'package:the_movie_db_app/widgets/cast_card_name.dart';
+
+export 'package:the_movie_db_app/widgets/image_card.dart';
+
+export 'package:the_movie_db_app/widgets/row_icon_details.dart';
+
 export 'package:the_movie_db_app/widgets/bottom_navbar_proyect.dart';
 
 export 'package:the_movie_db_app/widgets/carrousel_continue_waching.dart';
@@ -11,3 +17,19 @@ export 'package:the_movie_db_app/widgets/header_carrousel.dart';
 export 'package:the_movie_db_app/widgets/card_header_carrousel.dart';
 
 export 'package:the_movie_db_app/widgets/appbar_proyect.dart';
+
+export 'package:the_movie_db_app/widgets/cast_crew_card.dart';
+
+export 'package:the_movie_db_app/widgets/fake_navbar.dart';
+
+export 'package:the_movie_db_app/widgets/gradient_image.dart';
+
+export 'package:the_movie_db_app/widgets/header_appbar.dart';
+
+export 'package:the_movie_db_app/widgets/movie_description.dart';
+
+export 'package:the_movie_db_app/widgets/movie_info.dart';
+
+export 'package:the_movie_db_app/widgets/principal_image.dart';
+
+export 'package:the_movie_db_app/widgets/row_icon_header.dart';
